+++ src/api/tournaments.js
@@ -1,14 +1,15 @@
 import { axios } from '.';
 
 export const getTournaments = async () => {
-  const { data } = await axios.get(
-    'tournaments',
-    {
-      params: {
-        limit: 25,
-        offset: 0,
-      }
+  const id = "38fb2aac-544f-47a5-8658-6273936a9a2b";
+  const { data } = await axios({
+    method: 'get',
+    url: `tournaments/user/${id}`,
+    params: {
+      limit: 25,
+      offset: 0,
     }
+  }
   );
   return data;
 }
@@ -41,5 +42,7 @@ export const saveTournament = async ({
 }
 
 export const deleteTournament = async (id) => {
-  await axios.delete(`tournaments/${id}`);
+  if (confirm("Do you want to delete this Tournament?")) {
+    await axios.delete(`tournaments/${id}`);
+  }
 }
